Q124-001 disable all statistics except resourceUnitsUsed
Disable all statistics except resourceUnitsUsed to minimize the amount of
memory needed for caching the CVC4 outputs.

This commit is meant to be as non-intrusive as possible; in particular, we
not want to disable all existing statistics individually nor disable new
ones when they are added by the CVC4 team.

* src/util/statistics_registry.h
* src/util/statistics_registry.cpp
(registerStat): add an argument to ignore statistic.
(unregisterStat): same as above.

* src/smt/smt_engine.cpp
(SmtEngineStatistics): keep "resourceUnitsUsed" statistics registered.
(~SmtEngineStatistics): similar as above, needed to match the above change.

Change-Id: I18ac96ee55822f25f31add68b4912ba50906d87d